Moving ranks to their own folder
