+++ MekHQ/src/mekhq/campaign/MercRosterAccess.java
@@ -36,8 +36,8 @@ import megamek.common.UnitType;
 import mekhq.MekHQ;
 import mekhq.campaign.force.Force;
 import mekhq.campaign.personnel.Person;
-import mekhq.campaign.personnel.Rank;
-import mekhq.campaign.personnel.Ranks;
+import mekhq.campaign.personnel.ranks.Rank;
+import mekhq.campaign.personnel.ranks.Ranks;
 import mekhq.campaign.personnel.Skill;
 import mekhq.campaign.personnel.SkillType;
 import mekhq.campaign.unit.Unit;
