+++ MekHQ/src/mekhq/campaign/personnel/ranks/Rank.java
@@ -17,7 +17,7 @@
  * You should have received a copy of the GNU General Public License
  * along with MekHQ. If not, see <http://www.gnu.org/licenses/>.
  */
-package mekhq.campaign.personnel;
+package mekhq.campaign.personnel.ranks;
 
 import java.io.PrintWriter;
 import java.util.ArrayList;
