+++ MekHQ/src/mekhq/campaign/personnel/ranks/Ranks.java
@@ -19,7 +19,7 @@
  * You should have received a copy of the GNU General Public License
  * along with MekHQ. If not, see <http://www.gnu.org/licenses/>.
  */
-package mekhq.campaign.personnel;
+package mekhq.campaign.personnel.ranks;
 
 import java.io.FileInputStream;
 import java.io.InputStream;
