+++ MekHQ/src/mekhq/gui/dialog/NewRecruitDialog.java
@@ -27,10 +27,9 @@ import megamek.client.ui.swing.dialog.imageChooser.AbstractIconChooser;
 import megamek.client.ui.swing.dialog.imageChooser.PortraitChooser;
 import megamek.common.enums.Gender;
 import megamek.common.util.EncodeControl;
-import mekhq.MHQStaticDirectoryManager;
 import mekhq.MekHQ;
 import mekhq.campaign.personnel.Person;
-import mekhq.campaign.personnel.Ranks;
+import mekhq.campaign.personnel.ranks.Ranks;
 import mekhq.gui.CampaignGUI;
 import mekhq.gui.preferences.JWindowPreference;
 import mekhq.gui.view.PersonViewPanel;
